// src/types/profile.ts
export interface EmployeeData {
    first_name: string;
    last_name: string;
    email: string;
    phone: string;
    hire_date: string;
    avatar_url: string;
    location: string;
    department_id: number;
    employee_id: number;
    last_login_at: string;
    bg_img_url: string;
    role_name?: string;
}